+++ src/app/page.tsx
@@ -1,32 +1,20 @@
-"use client";
-
-import ImageSlider from "@/components/imageslider/ImageSlider";
-import MyProjectPosts from "@/components/myprojectpost/MyProjectPosts";
-import Posts from "@/components/posts/Posts";
-import Search from "@/components/posts/Search";
-import PostTabMenu from "@/components/posttabmenu/PostTabMenu";
-import { activeTabState } from "@/store/MainStateStore";
+import ImageSlider from "@/components/ui/ImageSlider";
+import Search from "@/components/main/posts/Search";
+import PostTabMenu from "@/components/main/postTabMenu/PostTabMenu";
 import React from "react";
-import { useRecoilValue } from "recoil";
+import PostTabContents from "@/components/main/PostTabContents";
 
 function HomePage() {
-  const currentTab = useRecoilValue(activeTabState);
-
-  return (
-    <>
-      <ImageSlider />
-      <div className="pc:max-w-[1200px] tablet:max-w-[750px] mobile:max-w-[340px] mx-auto mt-10">
-        {currentTab ? null : (
-          <div className="pc:hidden tablet:hidden">
-            <Search />
-          </div>
-        )}
-
-        <PostTabMenu />
-        {currentTab ? <MyProjectPosts /> : <Posts />}
-      </div>
-    </>
-  );
+    return (
+        <>
+            <ImageSlider/>
+            <div className="mt-10">
+                <Search/>
+                <PostTabMenu/>
+                <PostTabContents/>
+            </div>
+        </>
+    );
 }
 
 export default HomePage;
