+++ src/components/main/postTabMenu/PostTabMenu.tsx
@@ -1,29 +1,30 @@
+'use client';
 import React from "react";
 import { useRecoilState } from "recoil";
 import { activeTabState } from "@/store/MainStateStore";
 
 const PostTabMenu = () => {
-  const [activeTab, setActiveTab] = useRecoilState(activeTabState);
+  const [isMyProjectPostsTab, setIsMyProjectPostsTab] = useRecoilState(activeTabState);
 
   return (
     <div className="flex border-b mt-5">
       <div
         className={`p-5 font-bold text-2xl cursor-pointer mobile:text-xl ${
-          activeTab
+            isMyProjectPostsTab
             ? "text-greyUnselect"
             : "border-b-2 border-black100 text-black100"
         }`}
-        onClick={() => setActiveTab(false)}
+        onClick={() => setIsMyProjectPostsTab(false)}
       >
         게시글
       </div>
       <div
         className={`p-5 font-bold text-2xl cursor-pointer mobile:text-xl ${
-          activeTab
+            isMyProjectPostsTab
             ? "border-b-2 border-black100 text-black100"
             : "text-greyUnselect"
         }`}
-        onClick={() => setActiveTab(true)}
+        onClick={() => setIsMyProjectPostsTab(true)}
       >
         내 프로젝트
       </div>
