+++ src/components/main/posts/PositionDropdown.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Image from "next/image";
-import Arrow from "../../../public/images/bottomarrow.svg";
+import Arrow from "../../../../public/images/bottomarrow.svg";
 import { useRecoilValue } from "recoil";
 import PositionSelect from "./PositionSelect";
 import {
